Made the heatmap script the food script since the old one is not needed
