+++ unity/Group55TestProjekt/Assets/Scripts/Food.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 54ff8b54d9f1ea145ae06cf0c3cc1c94
+guid: f7e797ed434e3bc479870b94147dada9
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
