--- unity/Group55TestProjekt/Assets/Scripts/heatmap/Heatmap.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f7e797ed434e3bc479870b94147dada9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
